Auto generate top and side menu items
1. Top menu, uses a help to pick correct pages and build the menu in the layout file.
2. Side Menu. For the demo using 'User Guide', helper indexes the correct pages and
	uses a partial in the page to display the menu.

Added new helpers in the config.rb file to index and order the correct
pages to auto generate the menus.

Top menu in the layout file will look for `menu: top` and `menuindex: n`,
these will use `title` in the front matter as the menu item.

Side menu is used for the user guide pages, and inserts a partial `_user_nav.erb`,
This helper uses `category` and `order` to build the menu located in the partial.

Added headings to the page which inserts `heading` from the file.